+++ modules/unitframes/elements/debuffindicator.lua
@@ -57,7 +57,7 @@ local function frame_UpdateDebuffIndicator(self)
 end
 
 function UF:CreateDebuffIndicator(frame, textParent)
-	local element = (textParent or frame):CreateFontString(nil, "ARTWORK", "LSStatusIcon12Font")
+	local element = (textParent or frame):CreateFontString(nil, "ARTWORK", "LSIcon12Font")
 	element:SetJustifyH("CENTER")
 	element:SetJustifyV("MIDDLE")
 	element:SetNonSpaceWrap(true)