+++ modules/unitframes/units/focus.lua
@@ -122,7 +122,7 @@ function UF:CreateFocusFrame(frame)
 
 	frame.Auras = self:CreateAuras(frame, "focus")
 
-	local status = textParent:CreateFontString(nil, "ARTWORK", "LSStatusIcon16Font")
+	local status = textParent:CreateFontString(nil, "ARTWORK", "LSIcon16Font")
 	status:SetJustifyH("RIGHT")
 	status:SetPoint("RIGHT", frame, "BOTTOMRIGHT", -4, -1)
 	frame:Tag(status, "[ls:questicon][ls:sheepicon][ls:phaseicon][ls:leadericon][ls:lfdroleicon][ls:classicon]")